added support to filter twitter query based on the input
